+++ packages/stream_chat/lib/src/client/client.dart
@@ -29,7 +29,8 @@ import 'package:stream_core/stream_core.dart'
         WsEvent;
 import 'package:synchronized/synchronized.dart';
 
-import '../../open_api/api.dart' show DefaultApi, DurationResponse, ListDevicesResponse, SearchRolesResponse;
+import '../../open_api/api.dart'
+    show CreateDeviceRequestPushProvider, DefaultApi, ListDevicesResponse, SearchRolesResponse;
 import '../../version.dart';
 import '../core/api/attachment_file_uploader.dart';
 import '../core/api/requests.dart';
@@ -55,7 +56,6 @@ import '../core/models/poll.dart';
 import '../core/models/poll_option.dart';
 import '../core/models/poll_vote.dart';
 import '../core/models/push_preference.dart';
-import '../core/models/push_provider.dart';
 import '../core/models/reaction.dart';
 import '../core/models/role_type.dart';
 import '../core/models/thread.dart';
@@ -1208,9 +1208,9 @@ class StreamChatClient {
   ///
   /// [pushProviderName] names which of the app's configurations for
   /// [pushProvider] to use, for apps that have more than one.
-  Future<Result<DurationResponse>> addDevice(
+  Future<Result<void>> addDevice(
     String id,
-    PushProvider pushProvider, {
+    CreateDeviceRequestPushProvider pushProvider, {
     String? pushProviderName,
   }) => _devicesRepository.addDevice(
     id,
@@ -1222,7 +1222,7 @@ class StreamChatClient {
   Future<Result<ListDevicesResponse>> getDevices() => _devicesRepository.getDevices();
 
   /// Removes a registered device, stopping push notifications to it.
-  Future<Result<DurationResponse>> removeDevice(String id) => _devicesRepository.removeDevice(id);
+  Future<Result<void>> removeDevice(String id) => _devicesRepository.removeDevice(id);
 
   /// Set push preferences for the current user.
   ///